+++ komock-core/src/main/kotlin/ua/com/lavi/komock/engine/NamedQueuedThreadPool.kt
@@ -1,0 +1,14 @@
+package ua.com.lavi.komock.engine
+
+import org.eclipse.jetty.util.thread.QueuedThreadPool
+
+/**
+ * Created by Oleksandr Loushkin on 10.07.2017.
+ */
+class NamedQueuedThreadPool(maxThreads: Int, minThreads: Int, idleTimeout: Int, threadName: String) :
+        QueuedThreadPool(maxThreads, minThreads, idleTimeout) {
+
+    init {
+        name = threadName
+    }
+}